ADD: graphql error tests
move files around
